Postgresql migration arc
tried multithreading on sqlite, shat myself, and migrated to post

+++ backend/users.py
@@ -1,15 +1,18 @@
 #Imports
-from datetime import datetime, timezone, timedelta  
+from datetime import datetime, time, timezone, timedelta  
 from enum import Enum
+from sqlite3 import OperationalError
+from threading import Lock
 from typing import Any, Dict, List, Tuple
 from flask import Flask, render_template, session, Request, url_for
 import passwords
 import phonenumbers
+from sqlalchemy import func, type_coerce
 from phonenumbers import NumberParseException, PhoneNumber, PhoneNumberFormat
 from email_validator import EmailNotValidError, EmailSyntaxError, validate_email 
 from flask_login import UserMixin, login_user, current_user
 from flask_mail import Mail, Message
-from sqlalchemy.orm import sessionmaker, Session, Query
+from sqlalchemy.orm import sessionmaker, Session, Query, scoped_session
 import global_objects
 from global_objects import db, login_manager, Gender
 from sqlalchemy_utils import EmailType, CountryType, PhoneNumberType, Country
@@ -21,7 +24,9 @@ from itsdangerous import BadSignature, SignatureExpired, URLSafeTimedSerializer
 app : Flask
 
 sessionmaker_jobtaker : sessionmaker
+jobtaker_lock = Lock()
 sessionmaker_jobtaker : sessionmaker
+jobmaker_lock = Lock()
 
 def init_module(application):
     #initializes the users module
@@ -41,22 +46,22 @@ def user_model_factory(bind_key, _roles):
     class DynamicUser(db.Model, UserMixin):
         __bind_key__ = bind_key
         __tablename__ = bind_key
-        id = db.Column(db.Integer, primary_key=True, unique=True)
+        id = db.Column(db.Integer(), primary_key=True, unique=True)
         first_name = db.Column(db.String(30), nullable=False)
         last_name = db.Column(db.String(30), nullable=False)
         email = db.Column(EmailType, nullable=False, unique=True)
         phone_number = db.Column(PhoneNumberType, nullable=False, unique=True)
         country = db.Column(CountryType, nullable=False)
         city = db.Column(db.String(40), nullable=False)
-        birth_date = db.Column(db.DateTime, default=datetime.min, nullable=False)
-        time_registered = db.Column(db.DateTime, default=datetime.now(timezone.utc).astimezone(), nullable=False)
-        confirmation_email_id = db.Column(db.Integer, nullable=True)
-        time_confirmed = db.Column(db.DateTime, nullable=True)
-        rating_avg = db.Column(db.Integer, nullable=True)
-        rating_count = db.Column(db.Integer, nullable=True)
+        birth_date = db.Column(db.DateTime(), default=datetime.min, nullable=False)
+        time_registered = db.Column(db.DateTime(), default=datetime.now(timezone.utc).astimezone(), nullable=False)
+        confirmation_email_id = db.Column(db.Integer(), nullable=True)
+        time_confirmed = db.Column(db.DateTime(), nullable=True)
+        rating_avg = db.Column(db.Integer(), nullable=True)
+        rating_count = db.Column(db.Integer(), nullable=True)
         gender = db.Column(db.Enum(Gender), nullable=False)
         pronouns = db.Column(db.String(11), nullable=False)
-        job_ammount_done = db.Column(db.Integer, default=0, nullable=False)
+        job_ammount_done = db.Column(db.Integer(), default=0, nullable=False)
         roles = _roles
     
         def __init__(self, first_name, last_name, email, phone_number, country, city, birth_date, gender, pronouns):
@@ -72,7 +77,22 @@ def user_model_factory(bind_key, _roles):
 
         def __repr__(self) -> str:
             return f"""User{'\n'}{self.id}{'\n'}{self.first_name}{'\n'}{self.last_name}{'\n'}{self.email}{'\n'}{self.phone_number}{'\n'}{self.country}{'\n'}{self.city}{'\n'}
-        {self.birth_date}{'\n'}{self.time_registered}{'\n'}{self.confirmation_email_id}{'\n'}{self.time_confirmed}{'\n'}{self.rating_avg}{'\n'}{self.job_ammount_done}{'\n'}{self.gender}{'\n'}{self.pronouns}{'\n'}"""
+        {self.birth_date}{'\n'}{self.time_registered}{'\n'}{self.confirmation_email_id}{'\n'}tcon {self.time_confirmed}{'\n'}{self.rating_avg}{'\n'}{self.job_ammount_done}{'\n'}{self.gender}{'\n'}{self.pronouns}{'\n'}"""
+        
+        def revoke_registration(self, db_session : Session | None = None, commit : bool = True) -> bool:
+            if self == global_objects.admin:
+                return False
+            db_was_empty = db_session == None
+            if db_was_empty:
+                db_session = get_user_session(self.__bind_key__ == global_objects.JOBTAKERS_BINDKEY)
+                db_session.begin()
+            
+            self.time_confirmed = None
+            
+            if db_was_empty or commit:
+                db_session.commit()
+                db_session.close()
+            return True
         
         def confirm_registration(self, db_session : Session | None = None, commit : bool = True) -> bool:
             if self.time_confirmed != None:
@@ -83,9 +103,12 @@ def user_model_factory(bind_key, _roles):
             if db_was_empty:
                 db_session = get_user_session(self.__bind_key__ == global_objects.JOBTAKERS_BINDKEY)
                 db_session.begin()
+                
             self.time_confirmed = datetime.now(timezone.utc).astimezone()
+            
             if db_was_empty or commit:
                 db_session.commit()
+                db_session.close()
             return True
         
         def change_rating(self, rating : int, db_session : Session | None = None, commit : bool = True) -> bool:
@@ -103,6 +126,7 @@ def user_model_factory(bind_key, _roles):
                 self.rating_count += 1
             if db_was_empty or commit:
                 db_session.commit()
+                db_session.close()
             return True
         
         def increment_job_ammount_done(self, db_session : Session | None = None, commit : bool = True) -> bool:
@@ -115,6 +139,7 @@ def user_model_factory(bind_key, _roles):
             self.job_ammount_done += 1
             if db_was_empty or commit:
                 db_session.commit()
+                db_session.close()
             return True
             
         def edit_account(self, request : Request, db_session : Session | None = None, commit : bool = True) -> bool:#request.form should contain account data minus email and an old and new password
@@ -149,13 +174,16 @@ def user_model_factory(bind_key, _roles):
             db_session.flush()
             search = passwords.get_password_by_id(self.id, self.__bind_key__ == global_objects.JOBTAKERS_BINDKEY)
             if(search == None):
+                db_session.close()
                 return False
             (password_session, query) = search
             if not query.first().edit_password(request.form['password'], password_session):
+                db_session.close()
                 return False
             password_session.commit()
             if db_was_empty or commit:
                 db_session.commit()
+                db_session.close()
             print(self)
             return True
         
@@ -169,14 +197,17 @@ def user_model_factory(bind_key, _roles):
                 db_session.begin()
             search = passwords.get_password_by_id(self.id, self.__bind_key__ == global_objects.JOBTAKERS_BINDKEY)
             if(search == None):
+                db_session.close()
                 return False
             (password_session, query) = search
             if not query.first().delete_password(password_session):
+                db_session.close()
                 return False
             password_session.commit()
             db_session.delete(self)
             if db_was_empty or commit:
                 db_session.commit()
+                db_session.close()
             return True
         
     return DynamicUser
@@ -185,33 +216,40 @@ JobTaker = user_model_factory(global_objects.JOBTAKERS_BINDKEY, {global_objects.
 JobMaker = user_model_factory(global_objects.JOBMAKERS_BINDKEY, {global_objects.jobmaker_role})
 Admin = user_model_factory("", {global_objects.jobtaker_role, global_objects.jobmaker_role, global_objects.admin_role})
 
-def get_user_type(is_jobtaker):
+def get_user_type(is_jobtaker : bool):
     #returns the user type for jobtakers or jobmakers
     if is_jobtaker:
         return JobTaker
     else:
         return JobMaker
     
-def get_user_session(is_jobtaker) -> Session:
+def get_user_session(is_jobtaker : bool) -> Session:
     #returns a db session for jobtakers or jobmakers
     if is_jobtaker:
         return sessionmaker_jobtaker()
     else:
         return sessionmaker_jobmaker()
     
-def get_user_by_id(user_id, is_jobtaker) -> Tuple[Session, Query[Any]] | None:
+def get_session_lock(is_jobtaker : bool) -> Lock:
+    #returns a db session lock for jobtakers or jobmakers
+    if is_jobtaker:
+        return jobtaker_lock()
+    else:
+        return jobmaker_lock()
+
+def get_user_by_id(user_id : int, is_jobtaker : bool) -> Tuple[Session, Query[Any]] | None:
     #gets a user by id
     if(user_id == 0):
         return None
     db_session = get_user_session(is_jobtaker)
     db_session.begin()
     user_type = get_user_type(is_jobtaker)
-    query = db_session.query(user_type).filter(user_type.id == user_id)
+    query = db_session.query(user_type).filter(user_type.id == int(user_id))
     if(query.count() != 1):
         return None
     return (db_session, query)
 
-def get_user_by_email(email, is_jobtaker) -> Tuple[Session, Query[Any]] | None:
+def get_user_by_email(email : str, is_jobtaker : bool) -> Tuple[Session, Query[Any]] | None:
     #gets user by email
     try:
         validate_email(email, check_deliverability=False)
@@ -223,12 +261,12 @@ def get_user_by_email(email, is_jobtaker) -> Tuple[Session, Query[Any]] | None:
     db_session = get_user_session(is_jobtaker)
     db_session.begin()
     user_type = get_user_type(is_jobtaker)
-    query = db_session.query(user_type).filter(user_type.email == email)
+    query = db_session.query(user_type).filter(user_type.email == str(email))
     if(query.count() != 1):
         return None
     return (db_session, query)
 
-def get_user_by_phone(phone_number, is_jobtaker) -> Tuple[Session, Query[Any]] | None:
+def get_user_by_phone(phone_number : str | PhoneNumber, is_jobtaker : bool) -> Tuple[Session, Query[Any]] | None:
     #gets user by phone number
     if(not isinstance(phone_number, PhoneNumber)):
         phone_number = phonenumbers.parse(phone_number)
@@ -251,13 +289,17 @@ def get_user_by_phone(phone_number, is_jobtaker) -> Tuple[Session, Query[Any]] |
 @login_manager.user_loader
 def load_user(user_id):
     #gets a user depending on the session account type
-    
     if(session.get('account_type') == global_objects.ADMIN_SESSION_NAME):
         return global_objects.admin
     elif(session.get('account_type') == global_objects.JOBTAKER_SESSION_NAME or session.get('account_type') == global_objects.JOBMAKER_SESSION_NAME):
         db_session = get_user_session(session.get('account_type') == global_objects.JOBTAKER_SESSION_NAME)
         db_session.begin()
-        return db_session.query(get_user_type(session.get('account_type') == global_objects.JOBTAKER_SESSION_NAME)).get({"id":int(user_id)})
+        search = get_user_by_id(user_id, session.get('account_type') == global_objects.JOBTAKER_SESSION_NAME)
+        if(search == None):
+            return None
+        user = search[1].first()
+        db_session.close()
+        return user
     else:
         return None
 
@@ -316,16 +358,19 @@ def verify_login_attempt(request : Request, is_jobtaker : bool) -> bool:#request
     search = get_user_by_email(request.form['email'], is_jobtaker)
     if(search == None):
         return False
-    (_, user_query) = search 
-    search = passwords.get_password_by_id(user_query.first().id, is_jobtaker)
+    (user_session, user) = (search[0], search[1].first())
+    user_session.close()
+    search = passwords.get_password_by_id(user.id, is_jobtaker)
     if(search == None):
+        password_session.close()
         return False
-    (_, password_query) = search 
-    if not password_query.first().check_password_hash(request.form['password']):
+    (password_session, password) = (search[0], search[1].first())
+    password_session.close()
+    if not password.check_password_hash(request.form['password']):
         return False
     session['account_type'] = (lambda x : global_objects.JOBTAKER_SESSION_NAME if x else global_objects.JOBMAKER_SESSION_NAME)(is_jobtaker)
-    login_user(user_query.first(), remember=True)
-    identity_changed.send(app, identity=Identity(user_query.first().id))
+    login_user(user, remember=True)
+    identity_changed.send(app, identity=Identity(user.id))
     return True
 
 def verify_admin_login_attempt(password) -> bool:
@@ -358,27 +403,37 @@ def register_user_account(request : Request, is_jobtaker : bool, confirmation_em
     user_session = get_user_session(is_jobtaker)
     user_session.begin()
     if get_user_by_email(user.email, is_jobtaker) != None:
+        user_session.close()
         return False
     if get_user_by_phone(user.phone_number, is_jobtaker) != None:
+        user_session.close()
         return False
     user_session.add(user)
     user_session.flush()
     user_id = user.id
-    if not passwords.add_user_password_to_db(user_id, request.form['password'], is_jobtaker):
+    if not passwords.add_user_password_to_db(user_id, request.form['password'], is_jobtaker): 
+        user_session.close()
         return False
     user_session.commit()#we commit it after adding the password, to avoid desync in case of failure to commit password
     session['account_type'] = global_objects.JOBTAKER_SESSION_NAME if is_jobtaker else global_objects.JOBMAKER_SESSION_NAME
     result = send_confirmation_email(user.id, is_jobtaker)
+    user_session.close()
     return result
 
-def send_confirmation_email(user_id : int, is_jobtaker : bool) -> bool:#expects user to exist in db
+def send_confirmation_email(user_id : int, is_jobtaker : bool) -> bool | str:#expects user to exist in db
     (db_session, query) = get_user_by_id(user_id, is_jobtaker)
     user = query.first()
     user.confirmation_email_id = 0 if user.confirmation_email_id is None else user.confirmation_email_id + 1
     if user.confirmation_email_id >= 5: #you wont need more than 5 conf emails
+        db_session.close()
         return False
     serializer = URLSafeTimedSerializer(app.config['SECRET_KEY'])
     token = serializer.dumps({"id" : user.id, "is_jobtaker" : is_jobtaker, "confirmation_email_id" : user.confirmation_email_id}, salt=global_objects.EMAIL_CONFIRMATION_SALT)
+    
+    if global_objects.testing:
+        db_session.commit()
+        db_session.close()
+        return url_for('confirm_account', token=token, _external = True)
     
     mail = Mail(app)
     msg = Message(subject="Teenquest Email Confirmation", sender=app.config["MAIL_USERNAME"], recipients=[user.email])
@@ -386,6 +441,7 @@ def send_confirmation_email(user_id : int, is_jobtaker : bool) -> bool:#expects 
     msg.html = render_template("/email_templates/confirmation_email.html", confirmation_url=url)
     mail.send(msg)
     db_session.commit()
+    db_session.close()
     return True
 
 def confirm_user_account(token : str, expiration_time : int = 3600) -> Tuple[Any, bool] | int:
@@ -401,14 +457,19 @@ def confirm_user_account(token : str, expiration_time : int = 3600) -> Tuple[Any
         print(user_data)
         search = get_user_by_id(user_data["id"], user_data["is_jobtaker"])
         if search is None:
+            db_session.close()
             return 0
         (db_session, query) = search
         if query.first().confirmation_email_id != user_data["confirmation_email_id"]:
+            db_session.close()
             return 1
         result = query.first().confirm_registration(db_session)
         if not result:
+            db_session.close()
             return 2
-        return (query.first(), user_data["is_jobtaker"]) #this is the most pythonic shit ever
+        user = query.first()
+        db_session.close()
+        return (user, user_data["is_jobtaker"]) #this is the most pythonic shit ever
     except SignatureExpired:
         return 3
     except BadSignature:
@@ -416,17 +477,22 @@ def confirm_user_account(token : str, expiration_time : int = 3600) -> Tuple[Any
     
 def delete_unconfirmed_users() -> bool:
     print("deleting unconf users")
-    i = 0
-    while i < 2:#i = 0 checks jobmakers and i = 1 checks jobtakers
-        user_type = get_user_type(bool(i))
-        db_session = get_user_session(bool(i))
-        query = db_session.query(user_type).filter(
-            user_type.time_confirmed is None,
-            datetime.now(timezone.utc).astimezone() - user_type.time_registered > timedelta(days=3))
-        print(f"wiped {query.count()} {"jobtakers" if bool(i) else "jobmakers"}")
-        query.delete()
-        db_session.commit()
-        i += 1
+    with app.app_context():
+        i = 0
+        while i < 2:#i = 0 checks jobmakers and i = 1 checks jobtakers
+            user_type = get_user_type(bool(i))
+            db_session = get_user_session(bool(i))
+            db_session.begin()
+            query = db_session.query(user_type.id).filter(
+                user_type.time_confirmed == None,
+                user_type.time_registered < (datetime.now().astimezone() - timedelta(days=3)))
+            user_ids : List[int] = [row.id for row in query.all()]
+            db_session.query(user_type).filter(user_type.id.in_(user_ids)).delete()
+            db_session.commit()
+            passwords.delete_passwords_by_ids(user_ids, bool(i))
+            print(f"wiped {len(user_ids)} {"jobtakers" if bool(i) else "jobmakers"}")
+            i += 1
+            db_session.close()
     return True
 
 @identity_loaded.connect
